+++ termbox-bindings-hs/src/Termbox/Bindings/Hs/Internal/EventType.hs
@@ -9,12 +9,12 @@ module Termbox.Bindings.Hs.Internal.EventType
 where
 
 import Data.Word (Word8)
-import qualified Termbox.Bindings.C
+import Termbox.Bindings.C (_TB_EVENT_KEY, _TB_EVENT_MOUSE, _TB_EVENT_RESIZE)
 
 -- | An event type.
 newtype Tb_event_type
   = Tb_event_type Word8
-  deriving stock (Eq, Ord)
+  deriving stock (Eq)
 
 instance Show Tb_event_type where
   show = \case
@@ -24,21 +24,21 @@ instance Show Tb_event_type where
 
 pattern TB_EVENT_KEY :: Tb_event_type
 pattern TB_EVENT_KEY <-
-  ((== Tb_event_type Termbox.Bindings.C._TB_EVENT_KEY) -> True)
+  ((== Tb_event_type _TB_EVENT_KEY) -> True)
   where
-    TB_EVENT_KEY = Tb_event_type Termbox.Bindings.C._TB_EVENT_KEY
+    TB_EVENT_KEY = Tb_event_type _TB_EVENT_KEY
 
 pattern TB_EVENT_MOUSE :: Tb_event_type
 pattern TB_EVENT_MOUSE <-
-  ((== Tb_event_type Termbox.Bindings.C._TB_EVENT_MOUSE) -> True)
+  ((== Tb_event_type _TB_EVENT_MOUSE) -> True)
   where
-    TB_EVENT_MOUSE = Tb_event_type Termbox.Bindings.C._TB_EVENT_MOUSE
+    TB_EVENT_MOUSE = Tb_event_type _TB_EVENT_MOUSE
 
 pattern TB_EVENT_RESIZE :: Tb_event_type
 pattern TB_EVENT_RESIZE <-
-  ((== Tb_event_type Termbox.Bindings.C._TB_EVENT_RESIZE) -> True)
+  ((== Tb_event_type _TB_EVENT_RESIZE) -> True)
   where
-    TB_EVENT_RESIZE = Tb_event_type Termbox.Bindings.C._TB_EVENT_RESIZE
+    TB_EVENT_RESIZE = Tb_event_type _TB_EVENT_RESIZE
 
 -- N.B. This requires Tb_event_type to remain abstract
 {-# COMPLETE TB_EVENT_KEY, TB_EVENT_MOUSE, TB_EVENT_RESIZE #-}